+++ spec/api-browser-window-spec.js
@@ -113,126 +113,6 @@ describe('BrowserWindow module', () => {
 
   afterEach(closeTheWindow)
 
-  describe('BrowserWindow.setAutoHideCursor(autoHide)', () => {
-    describe('on macOS', () => {
-      before(function () {
-        if (process.platform !== 'darwin') {
-          this.skip()
-        }
-      })
-
-      it('allows changing cursor auto-hiding', () => {
-        expect(() => {
-          w.setAutoHideCursor(false)
-          w.setAutoHideCursor(true)
-        }).to.not.throw()
-      })
-    })
-
-    describe('on non-macOS platforms', () => {
-      before(function () {
-        if (process.platform === 'darwin') {
-          this.skip()
-        }
-      })
-
-      it('is not available', () => {
-        expect(w.setAutoHideCursor).to.be.undefined()
-      })
-    })
-  })
-
-  describe('BrowserWindow.setWindowButtonVisibility()', () => {
-    before(function () {
-      if (process.platform !== 'darwin') {
-        this.skip()
-      }
-    })
-
-    it('does not throw', () => {
-      expect(() => {
-        w.setWindowButtonVisibility(true)
-        w.setWindowButtonVisibility(false)
-      }).to.not.throw()
-    })
-
-    it('throws with custom title bar buttons', () => {
-      expect(() => {
-        w.destroy()
-        w = new BrowserWindow({
-          show: false,
-          titleBarStyle: 'customButtonsOnHover',
-          frame: false
-        })
-        w.setWindowButtonVisibility(true)
-      }).to.throw('Not supported for this window')
-    })
-  })
-
-  describe('BrowserWindow.setVibrancy(type)', () => {
-    it('allows setting, changing, and removing the vibrancy', () => {
-      expect(() => {
-        w.setVibrancy('light')
-        w.setVibrancy('dark')
-        w.setVibrancy(null)
-        w.setVibrancy('ultra-dark')
-        w.setVibrancy('')
-      }).to.not.throw()
-    })
-  })
-
-  describe('BrowserWindow.setAppDetails(options)', () => {
-    before(function () {
-      if (process.platform !== 'win32') {
-        this.skip()
-      }
-    })
-
-    it('supports setting the app details', () => {
-      const iconPath = path.join(fixtures, 'assets', 'icon.ico')
-
-      expect(() => {
-        w.setAppDetails({ appId: 'my.app.id' })
-        w.setAppDetails({ appIconPath: iconPath, appIconIndex: 0 })
-        w.setAppDetails({ appIconPath: iconPath })
-        w.setAppDetails({ relaunchCommand: 'my-app.exe arg1 arg2', relaunchDisplayName: 'My app name' })
-        w.setAppDetails({ relaunchCommand: 'my-app.exe arg1 arg2' })
-        w.setAppDetails({ relaunchDisplayName: 'My app name' })
-        w.setAppDetails({
-          appId: 'my.app.id',
-          appIconPath: iconPath,
-          appIconIndex: 0,
-          relaunchCommand: 'my-app.exe arg1 arg2',
-          relaunchDisplayName: 'My app name'
-        })
-        w.setAppDetails({})
-      }).to.not.throw()
-
-      expect(() => {
-        w.setAppDetails()
-      }).to.throw('Insufficient number of arguments.')
-    })
-  })
-
-  describe('BrowserWindow.fromId(id)', () => {
-    it('returns the window with id', () => {
-      expect(BrowserWindow.fromId(w.id).id).to.equal(w.id)
-    })
-  })
-
-  describe('BrowserWindow.fromWebContents(webContents)', () => {
-    let contents = null
-
-    beforeEach(() => { contents = webContents.create({}) })
-
-    afterEach(() => { contents.destroy() })
-
-    it('returns the window with the webContents', () => {
-      expect(BrowserWindow.fromWebContents(w.webContents).id).to.equal(w.id)
-      expect(BrowserWindow.fromWebContents(contents)).to.be.undefined()
-    })
-  })
-
   describe('BrowserWindow.fromDevToolsWebContents(webContents)', () => {
     let contents = null
 
@@ -248,14 +128,6 @@ describe('BrowserWindow module', () => {
         done()
       })
       w.webContents.openDevTools()
-    })
-  })
-
-  describe('BrowserWindow.openDevTools()', () => {
-    it('does not crash for frameless window', () => {
-      w.destroy()
-      w = new BrowserWindow({ show: false })
-      w.openDevTools()
     })
   })
